+++ src/pages/Account/Security.vue
@@ -8,22 +8,15 @@
     <ion-content padding>
         <h1> SECURITY </h1>
     </ion-content>
+    <BottomNav />
   </ion-page>
 </template>
 
 <script>
-
-import { add } from "ionicons/icons";
-import { addIcons } from "ionicons";
-addIcons({
-  "ios-add": add.ios,
-  "md-add": add.md
-});
+import BottomNav from '../../components/BottomNav'
 export default {
   name: "Security",
-  props: {
-    msg: String
-  },
+  components: { BottomNav },
   data() {
       return {