+++ src/pages/Droplets/DropletStatus.vue
@@ -8,28 +8,22 @@
     <ion-content padding>
         <h1> INDIVIDUAL DROPLET STATUS </h1>
     </ion-content>
+    <BottomNav />
   </ion-page>
 </template>
 
 <script>
+import BottomNav from '../../components/BottomNav'
 
-import { add } from "ionicons/icons";
-import { addIcons } from "ionicons";
-addIcons({
-  "ios-add": add.ios,
-  "md-add": add.md
-});
 export default {
-  name: "replace-this",
-  props: {
-    msg: String
-  },
+  name: "DropletStatus",
+  components: { BottomNav },
   data() {
-      return {
+        return {
         
-      }
-  }
-};
+        }
+    }
+}
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->