+++ changelog.md
@@ -1,3 +1,11 @@
+## 1/6/2020
+```
+- Renamed the 'name' field in each component's export default field
+- Removed unnecessary 'props' and 'data' fields in export default for each component and page
+- Added BottomNav component for each page to quickly get back to the dashboard
+- Added IonIcons successfully to BottomNav component
+```
+
 ## 31/5/2020
 ```
 - Created basic bottom nav bar, added dashboard link to nav